Fix progress overlay size
By not having AztecMediaSpan always adjust the bounds of the input drawable. When the input drawable is a LayerDrawable, setting its bounds messes with the individual layer bounds and in this case the progress indication drawable gets broken.

So instead, have the called decide if the bounds should be adjusted by the AztecMediaSpan.