Added isInit flag to only calculate initial content hash when isInit is true
This helps preventing a new hash being calculated on copy/paste operations, marking the content
as NO_CHANGE when some new piece of text has just been pasted / inserted.
For context, fromHtml() was calculating the initial hash under the assumption that this method
was only being called when starting AztecText (or moving from HTML to the visual editing view,
which is the same to all effects).