Build: Update kotlin to 1.6.10
This update is necessary because otherwise, with newer versions of Java
(see Java 17), this project cannot be compiled due to the fact that with
JDK17 it is no longer possible to make a field accessible using
reflection. It seems that Kotlin 1.4.20 is doing it, rightfully so, as
that point of time this 1.4.20 version of Kotlin wasn't ready for
Java 17.

FYI: The code fails to compile with the below exception:

------------------------------------------------------------------------

> Task :aztec:compileDebugKotlin FAILED
Could not perform incremental compilation: Could not connect to Kotlin
compile daemon
Could not connect to kotlin daemon. Using fallback strategy.
exception: java.lang.ExceptionInInitializerError
        at org.jetbrains.kotlin.com.intellij.pom.java.LanguageLevel.
            <clinit>(LanguageLevel.java:25)
        ...
Caused by: java.lang.reflect.InaccessibleObjectException: Unable to make
protected void java.util.ResourceBundle.setParent
(java.util.ResourceBundle) accessible: module java.base does not "opens
java.util" to unnamed module @41e77c24
        at java.base/java.lang.reflect.AccessibleObject.
            checkCanSetAccessible(AccessibleObject.java:354)
        ...
        ... 19 more

------------------------------------------------------------------------

Also, as part of this update the 'max/mindBy' were replaced by
'max/mindByOrNull', but other than that everything else remained the
same.