Test: Document test failures due to expected actual assertion mismatch
Those test, and more so, their specific 'expected' outcome are being
documented so that when they get fixed, it becomes evident what was
changed and why.

This is done in order to make it easier to reason about what went wrong
and whether main source code needs to be updated as well, that is, since
there was no other explicit change but the 'minSdkVersion' to '24'
update that caused those failures.