Test: Fix edit text to html related test failures
Some of the failures were caused due to the '<em>' and '<strong>'
needing to be the other way around. Other failures were caused due to
the '<i>' and '<s>' needing to be the other way around. While, some more
failure were caused due to the '<a href="xyz">', '<b>' and/or '<s>'
needing to be the other way around.

This test failures are most probably related to Robolectric, and the
fact that the '@Config(sdk = [23])' had to be removed, thus implicitly
updated, and all that, due to the 'minSdkVersion' to '24' change.

However, there is also a slight possibility that this is a breaking
change due to the 'minSdkVersion' to '24' update. Thus, this change need
to be verified and tested.