Docs: Remove supports android info from readme
With this change the 'Supports Android 4.1+ (API 16 - Jelly Bean)' information is removed from the 'README.md' documentation as this information has been outdated for years now and it has the ability to become outdated with every year passing by. Users of this library should refer to the 'minSdkVersion' within the root level 'build.gradle' file to understand what's the latest min SDK version this library supports.